+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 2.12.0
 * Amplitude scaling functions duplicate water reference amplitudes when
 sensible.
+* Added hline option to the mrs_data plot function.
 
 # spant 2.11.0
 * Bug fix for Siemens RDA slice thickness detection.